griliches
Source: R/griliches.R

Hayashi Source: Blackburn, M., and D. Neumark, 1992, "Unobserved Ability, Efficiency Wages, and Interindustry Wage Differentials," Quarterly Journal of Economics, 107, 142 1-1436.

data('griliches')

Format

A data.frame with 758 observations on 20 variables:

southern_residence: = 1 if individual lives in southern state when first interviewed

southern_residence_80: = 1 individual lives in south in 1980

married: = 1 if individual is married when first interviewed

married_80: = 1 if individual is married when interviewed in 1980

lives_metro: = 1 if individual lives in metropolitan area when first interviewed

lives_metro_80: = 1 if individual lives in metropolitan area in 1980

mothers_educ: mother's education measured in years

iq_score: score on the IQ test

kww_score: score on "Knowlegde of the World of Work" test

year: year was first interviewed

age: age of individual when first interviewed

age_80: age of individual in 1980

education: years of completed schooling when first interviewed

education_80: years of completed schooling in 1980

experience: years of work experience when first interviewed

experience_80: years of work experience in 1980

tenure: years of tenure when first interviewed

tenure_80: years of tenure in 1980

log_wage: log wage when first interviewd

log_wage_80: log wage in 1980

Source

https://sites.google.com/site/fumiohayashi/hayashi-econometrics/data-for-empirical

Details

A panel data extract of the Young Men's Cohort of the National Longitudinal Survey (NLS-Y).

Notes

Used in Chapter 3.9 and the Empirical Exercise of Chapter 3. Data is the Blackburn - Neumark sample after eliminating data with missing mother's education.

Examples

 str(griliches)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	758 obs. of  20 variables:
#>  $ southern_residence   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ southern_residence_80: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ married              : num  0 0 0 0 1 0 1 1 1 1 ...
#>  $ married_80           : num  1 1 1 1 1 0 1 1 1 1 ...
#>  $ lives_metro          : num  1 1 1 1 1 1 1 0 1 0 ...
#>  $ lives_metro_80       : num  1 1 1 1 1 1 1 0 1 0 ...
#>  $ mothers_educ         : num  8 14 14 12 6 8 8 14 12 13 ...
#>  $ iq_score             : num  93 119 108 96 74 91 114 111 95 132 ...
#>  $ kww_score            : num  35 41 46 32 27 24 50 37 44 44 ...
#>  $ year                 : num  68 66 67 66 73 66 73 67 66 73 ...
#>  $ age                  : num  19 23 20 18 26 16 30 23 22 30 ...
#>  $ age_80               : num  31 37 33 32 34 30 38 36 36 38 ...
#>  $ education            : num  12 16 14 12 9 9 18 15 12 18 ...
#>  $ education_80         : num  12 18 14 12 11 10 18 15 12 18 ...
#>  $ experience           : num  0.462 0 0.423 0.333 9.013 ...
#>  $ experience_80        : num  10.6 11.4 11 13.1 14.4 ...
#>  $ tenure               : num  0 2 1 1 3 1 6 1 2 5 ...
#>  $ tenure_80            : num  2 16 9 7 5 0 14 1 16 13 ...
#>  $ log_wage             : num  5.9 5.44 5.71 5.48 5.93 ...
#>  $ log_wage_80          : num  6.64 6.69 6.71 6.48 6.33 ...