T-Bill and Inflation Rates in the United States.
Source: R/mishkin.R

Hayashi Source: Mishkin, F., 1992, "Is the Fisher Effect for Real?'Journal of Monetary Economics, 30, 195-215.

data('mishkin')

Format

A data.frame with 491 observations on 7 variables:

year: year

month: month

inflation_1: one month inflation rate, in percent, annualized. Uses a rental equivalence measure to incorporate residential housing

inflation_3: one month inflation rate, in percent, annualized. Uses a rental equivalence measure to incorporate residential housing

tbill_1: one month treasury bill rate, in percent, annualized

tbill_3: three month treasury bill rate, in percent, annualized

cpi: the consumer price index for urban consumers using all goods

Source

https://sites.google.com/site/fumiohayashi/hayashi-econometrics/data-for-empirical

Details

A time series data set at monthly frequency of the one-month T-bill rate and the monthly inflation, stated in percent at annual rates.

Notes

Used in Section 11 of Chapter 2 and Empirical Exercise 1 of Chapter 2.

Examples

 str(mishkin)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	491 obs. of  7 variables:
#>  $ year       : num  1950 1950 1950 1950 1950 1950 1950 1950 1950 1950 ...
#>  $ month      : num  2 3 4 5 6 7 8 9 10 11 ...
#>  $ inflation_1: num  -3.55 5.25 1.69 5.06 6.72 ...
#>  $ inflation_3: num  1.13 4 4.49 7.82 9.43 ...
#>  $ tbill_1    : num  1.1 1.13 1.12 1.15 1.16 ...
#>  $ tbill_3    : num  1.13 1.14 1.14 1.18 1.17 ...
#>  $ cpi        : num  23.5 23.6 23.6 23.7 23.8 24.1 24.3 24.4 24.6 24.7 ...