moneydemand
Source: R/moneydemand.R

Hayashi Source: Stock, J., and M. Watson, 1993, "A Simple Estimator of Cointegrating Vectors in Higher Order Integrated Systems," Econometrics, 61, 783-820.

data('moneydemand')

Format

A data.frame with 90 observations on 4 variables:

log_m1: natural log of M1 money supply

log_p: natural log of net national product (NNP) price deflator

log_nnp: natural log of net national product

interest_rate: commercial paper rate, annualized in percent

Source

https://sites.google.com/site/fumiohayashi/hayashi-econometrics/data-for-empirical

Details

A time series data set at annual frequency of US money demand. Data period is 1900 to 1989.

Notes

Used in Chapter 10.5 and the Empirical Exercise of Chapter 10.

Examples

 str(moneydemand)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	90 obs. of  4 variables:
#>  $ log_m1       : num  1.72 1.86 1.94 1.98 2.03 ...
#>  $ log_p        : num  2.09 2.09 2.12 2.13 2.15 ...
#>  $ log_nnp      : num  0.903 1.013 1.011 1.055 1.033 ...
#>  $ interest_rate: num  4.38 4.28 4.92 5.47 4.2 4.4 5.68 6.34 4.37 3.98 ...