Electricity Supply Data from 1955.
Source: R/nerlove.R

Hayashi Source: Nerlove, M., 1963, “Returns to Scale in Electricity Supply,” in C. Christ (ed.), Measurement in Economics: Studies in Mathematical Economics and Econometrics in Memory of Yehuda Grunfeld, Stanford: Stanford University Press.

data('nerlove')

Format

A data.frame with 145 observations on 5 variables:

total_cost: total cost of supply in millions of dollars

output: in billions of kilowatt hours

labor_price: the price of labor

fuel_price: the price of fuels

capital_price: the price of capital

Source

https://sites.google.com/site/fumiohayashi/hayashi-econometrics/data-for-empirical

Details

A cross-sectional data set on 145 firms in 44 states in the year 1955 with data on total costs, factor prices (the wage rate, the price of fuel, and the rental price of capital), and output of power plants.

Notes

Used in Section 7 of Chapter 1, the Empirical Exercise of Chapter 1, Empirical Exercise 2 in Chapter 2.

Examples

 str(nerlove)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	145 obs. of  5 variables:
#>  $ total_cost   : num  0.082 0.661 0.99 0.315 0.197 0.098 0.949 0.675 0.525 0.501 ...
#>  $ output       : num  2 3 4 4 5 9 11 13 13 22 ...
#>  $ price_labor  : num  2.09 2.05 2.05 1.83 2.12 2.12 1.98 2.05 2.19 1.72 ...
#>  $ price_fuel   : num  17.9 35.1 35.1 32.2 28.6 28.6 35.5 35.1 29.1 15 ...
#>  $ price_capital: num  183 174 171 166 233 195 206 150 155 188 ...