USD/GBP Exchange Rates over time.

Hayashi Source: Bekaert, G., and R. Hodrick, 1993, "On Biases in the Measurement of Foreign Exchange Risk Premiums," Journal of International Money and Finance, 12, 115-138.

data('pound')

Format

A data.frame with 778 observations on 4 variables:

date: date of the observation

spot_rate: the ask price of the US Dollar in units of the British Pound in the spot market on Friday of the current week,

forward_30: the ask price of the US Dollar in units of the British Pound in the 30-day forward market on Friday of the current week

spot_30: : the bid price of the US Dollar in units of the British Pound in the spot market on the delivery date on a current forward contract

Source

https://sites.google.com/site/fumiohayashi/hayashi-econometrics/data-for-empirical

Details

A time series data set at weekly frequency of the US Dollar / British Pound exchange rate. Data period is January 1975 to November 1989.

Notes

Used in the Empirical Exercise of Chapter 6.

Examples

 str(pound)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	778 obs. of  4 variables:
#>  $ date      : Date, format: "1975-01-03" "1975-01-10" ...
#>  $ spot_rate : num  0.427 0.425 0.425 0.419 0.42 ...
#>  $ forward_30: num  0.43 0.427 0.426 0.421 0.422 ...
#>  $ spot_30   : num  0.422 0.42 0.417 0.414 0.412 ...